+++ src/archetype/createProject/types.ts
@@ -1,9 +1,11 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 
+import { MavenProject } from "../../explorer/model/MavenProject";
 import { Archetype } from "../Archetype";
 
 export interface IProjectCreationMetadata {
+    title: string; // Project creation title name
     archetype?: Archetype; // temporary cached data between steps, used to select versions
     archetypeArtifactId?: string;
     archetypeGroupId?: string;
@@ -12,6 +14,7 @@ export interface IProjectCreationMetadata {
     artifactId?: string;
     targetFolder?: string;
     targetFolderHint?: string; // default value for folder picker dialog
+    parentProject?: MavenProject;
 }
 
 export interface IProjectCreationStep {